added engines to profiling as well as engine_test in test_profile (#532)
* added engines to profiling as well as engine_test in test_profile

* changed test_profile.py to flake8 standard

* set engine==SingleCoreEngine() as default

* Update pypesto/profile/task.py


* Update pypesto/profile/profile.py


* Update pypesto/profile/profile.py


* Update pypesto/profile/task.py


* Update test/profile/test_profile.py
