+++ pypesto/profile/profile.py
@@ -1,15 +1,14 @@
 import logging
-import numpy as np
 from typing import Callable, Union, Iterable
 
-from ..objective.constants import GRAD
-from ..optimize import Optimizer, OptimizerResult
+from ..engine import Engine, SingleCoreEngine
+from ..optimize import Optimizer
 from ..problem import Problem
 from ..result import Result
-from .result import ProfilerResult
 from .profile_next_guess import next_guess
 from .options import ProfileOptions
 from .util import initialize_profile
+from .task import ProfilerTask
 
 logger = logging.getLogger(__name__)
 
@@ -18,6 +17,7 @@ def parameter_profile(
         problem: Problem,
         result: Result,
         optimizer: Optimizer,
+        engine: Engine = None,
         profile_index: Iterable[int] = None,
         profile_list: int = None,
         result_index: int = 0,
@@ -38,6 +38,8 @@ def parameter_profile(
         The existence of an optimization result is obligatory.
     optimizer:
         The optimizer to be used along each profile.
+    engine:
+        The engine to be used.
     profile_index:
         List with the parameter indices to be profiled
         (by default all free indices).
@@ -86,139 +88,38 @@ def parameter_profile(
     # existing list of profiles
     global_opt = initialize_profile(problem, result, result_index,
                                     profile_index, profile_list)
+    # if engine==None set SingleCoreEngine() as default
+    if engine is None:
+        engine = SingleCoreEngine()
 
-    # loop over parameters for profiling
+    # create Tasks
+    tasks = []
+    # loop over parameters to create tasks
     for i_par in profile_index:
         # only compute profiles for free parameters
         if i_par in problem.x_fixed_indices:
             continue
 
-        # create an instance of ProfilerResult, which will be appended to the
-        # result object, when this profile is finished
         current_profile = result.profile_result.get_profiler_result(
             i_par=i_par, profile_list=profile_list)
 
-        # compute profile in descending and ascending direction
-        for par_direction in [-1, 1]:
-            # flip profile
-            current_profile.flip_profile()
+        task = ProfilerTask(
+            current_profile=current_profile,
+            problem=problem,
+            optimizer=optimizer,
+            options=profile_options,
+            create_next_guess=create_next_guess,
+            global_opt=global_opt,
+            i_par=i_par
+        )
+        tasks.append(task)
 
-            # compute the current profile
-            current_profile = walk_along_profile(
-                current_profile=current_profile,
-                problem=problem,
-                par_direction=par_direction,
-                optimizer=optimizer,
-                options=profile_options,
-                create_next_guess=create_next_guess,
-                global_opt=global_opt,
-                i_par=i_par)
+    # execute the tasks with Engine
+    indexed_profiles = engine.execute(tasks)
+
+    # fill in the ProfilerResults at the right index
+    for indexed_profile in indexed_profiles:
+        result.profile_result.list[-1][indexed_profile['index']] = \
+            indexed_profile['profile']
 
     return result
-
-
-def walk_along_profile(
-        current_profile: ProfilerResult,
-        problem: Problem,
-        par_direction: int,
-        optimizer: Optimizer,
-        options: ProfileOptions,
-        create_next_guess: Callable,
-        global_opt: float,
-        i_par: int
-) -> ProfilerResult:
-    """
-    This function computes half a profile, by walking ahead in positive
-    direction until some stopping criterion is fulfilled. A two-sided profile
-    is obtained by flipping the profile direction.
-
-    Parameters
-    ----------
-    current_profile:
-        The profile which should be computed
-    problem:
-        The problem to be solved.
-    par_direction:
-        Indicates profiling direction (+1, -1: ascending, descending)
-    optimizer:
-        The optimizer to be used along each profile.
-    global_opt:
-        log-posterior value of the global optimum
-    options:
-        Various options applied to the profile optimization.
-    create_next_guess:
-        Handle of the method which creates the next profile point proposal
-    i_par:
-        index for the current parameter
-
-    Returns
-    -------
-    current_profile:
-        The current profile, modified in-place.
-    """
-
-    # create variables which are needed during iteration
-    stop_profile = False
-
-    # while loop for profiling (will be exited by break command)
-    while True:
-        # get current position on the profile path
-        x_now = current_profile.x_path[:, -1]
-
-        # check if the next profile point needs to be computed
-        if par_direction == -1:
-            stop_profile = (x_now[i_par] <= problem.lb_full[[i_par]]) \
-                or (current_profile.ratio_path[-1] < options.ratio_min)
-
-        if par_direction == 1:
-            stop_profile = (x_now[i_par] >= problem.ub_full[[i_par]]) \
-                or (current_profile.ratio_path[-1] < options.ratio_min)
-
-        if stop_profile:
-            break
-
-        # compute the new start point for optimization
-        x_next = create_next_guess(x_now, i_par, par_direction,
-                                   options, current_profile, problem,
-                                   global_opt)
-
-        # fix current profiling parameter to current value and set
-        # start point
-        problem.fix_parameters(i_par, x_next[i_par])
-        startpoint = np.array([x_next[i] for i in problem.x_free_indices])
-
-        # run optimization
-        # IMPORTANT: This optimization will need a proper exception
-        # handling (coming soon)
-        if startpoint.size > 0:
-            optimizer_result = optimizer.minimize(problem, startpoint, '0',
-                                                  allow_failed_starts=False)
-        else:
-            # if too many parameters are fixed, there is nothing to do ...
-            fval = problem.objective([])
-            optimizer_result = OptimizerResult(
-                id='0', x=np.array([]), fval=fval, n_fval=0, n_grad=0, n_res=0,
-                n_hess=0, n_sres=0, x0=np.array([]), fval0=fval, time=0)
-            optimizer_result.update_to_full(problem=problem)
-
-        if optimizer_result[GRAD] is not None:
-            gradnorm = np.linalg.norm(optimizer_result[GRAD][
-                                      problem.x_free_indices])
-        else:
-            gradnorm = None
-
-        current_profile.append_profile_point(
-            x=optimizer_result.x,
-            fval=optimizer_result.fval,
-            ratio=np.exp(global_opt - optimizer_result.fval),
-            gradnorm=gradnorm,
-            time=optimizer_result.time,
-            exitflag=optimizer_result.exitflag,
-            n_fval=optimizer_result.n_fval,
-            n_grad=optimizer_result.n_grad,
-            n_hess=optimizer_result.n_hess)
-
-    # free the profiling parameter again
-    problem.unfix_parameters(i_par)
-
-    return current_profile
